refs #329 - do not garbage collect entry points, default jvm view and gateway server.

+++ py4j-java/src/main/java/py4j/commands/MemoryCommand.java
@@ -36,6 +36,7 @@ import java.io.BufferedWriter;
 import java.io.IOException;
 import java.util.logging.Logger;
 
+import py4j.GatewayServer;
 import py4j.Protocol;
 import py4j.Py4JException;
 
@@ -67,7 +68,9 @@ public class MemoryCommand extends AbstractCommand {
 		// EoC
 		reader.readLine();
 
-		gateway.deleteObject(objectId);
+		if (objectId != Protocol.ENTRY_POINT_OBJECT_ID && objectId != Protocol.DEFAULT_JVM_OBJECT_ID && objectId != Protocol.GATEWAY_SERVER_ID) {
+			gateway.deleteObject(objectId);
+		}
 
 		return Protocol.getOutputVoidCommand();
 	}
